# docs/swagger.yaml
components:
  schemas:
    domain.Result:
      properties:
        output:
          type: string
      type: object
    handlers.ErrorResponse:
      properties:
        error:
          type: string
      type: object
    handlers.authRequest:
      properties:
        password:
          type: string
        username:
          type: string
      type: object
    handlers.loginResponse:
      properties:
        token:
          type: string
      type: object
    handlers.registerResponse:
      properties:
        message:
          type: string
      type: object
    handlers.taskCreateRequest:
      properties:
        code:
          type: string
        translator:
          type: string
      type: object
    handlers.taskCreateResponse:
      properties:
        task_id:
          type: string
      type: object
    handlers.taskResultResponse:
      properties:
        result:
          $ref: '#/components/schemas/domain.Result'
      type: object
    handlers.taskStatusResponse:
      properties:
        status:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http
externalDocs:
  description: ""
  url: ""
info:
  contact:
    email: veniamin@belyaev.work
    name: Veniamin Belyaev
  description: The foundation for a safe remote code execution service
  license:
    name: MIT
  title: HTTP REST API for a safe remote code execution service
  version: "1.0"
openapi: 3.1.0
paths:
  /login:
    post:
      description: Logging in as a user
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.authRequest'
                description: login credentials
                summary: request
        description: login credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.loginResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      summary: Login as a user
      tags:
      - auth
  /register:
    post:
      description: Registering a user
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.authRequest'
                description: registration credentials
                summary: request
        description: registration credentials
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.registerResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      summary: Register a user
      tags:
      - auth
  /result/{task_id}:
    get:
      description: Getting the result of a task by id. Only the owner of the task
        can access it
      parameters:
      - description: task id
        in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.taskResultResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get the result of a certain task
      tags:
      - task
  /status/{task_id}:
    get:
      description: Getting the status of a task by id. Only the owner of the task
        can access it
      parameters:
      - description: task id
        in: path
        name: task_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.taskStatusResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get the status of a certain task
      tags:
      - task
  /task:
    post:
      description: Creating a task owned by an authenticated user and queuing it for
        execution
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.taskCreateRequest'
                description: task submission
                summary: request
        description: task submission
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.taskCreateResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a task
      tags:
      - task
servers:
- url: /